# docs/phase6/API_ARCHITECTURE/openapi-specification.yaml
openapi: 3.0.3
info:
  title: Archon Phase 6 Authentication API
  description: |
    Comprehensive RESTful API for Archon Phase 6 authentication system with multi-agent support.
    
    This API provides secure authentication, authorization, and user management capabilities
    optimized for high-performance multi-agent orchestration with sub-50ms response times.
    
    ## Features
    - JWT-based authentication with multi-layered caching
    - Role-based access control (RBAC) for 22+ specialized agents
    - Session management with Redis optimization
    - Rate limiting and security monitoring
    - Real-time event notifications via WebSocket
    - Comprehensive audit logging
    
    ## Performance Targets
    - Authentication latency: <50ms (p95), <200ms (p99)
    - Throughput: 10,000+ RPS
    - Cache hit rate: >90%
    - Uptime: 99.9%
    
  version: 2.0.0
  contact:
    name: Archon Development Team
    email: dev@archon.ai
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.archon.ai/v2
    description: Production server
  - url: https://staging-api.archon.ai/v2
    description: Staging server
  - url: http://localhost:8181/api/v2
    description: Local development server

# Security Schemes
security:
  - BearerAuth: []
  - ApiKeyAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authenticated requests
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for service-to-service communication
  schemas:
...
    LoginRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          description: User email address
        password:
          type: string
          minLength: 8
          description: User password
        remember_me:
          type: boolean
          default: false
          description: Extend session duration
        device_info:
          $ref: '#/components/schemas/DeviceInfo'
    DeviceInfo:
      type: object
      properties:
        device_id:
          type: string
          description: Unique device identifier
        user_agent:
          type: string
          description: Browser/client user agent
        ip_address:
          type: string
          format: ipv4
          description: Client IP address
        location:
          type: string
          description: Geographic location
...
    APIError:
      type: object
      required:
        - success
        - error
      properties:
        success:
          type: boolean
          enum: [false]
        data:
          type: object
          nullable: true
        error:
          type: object
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - INVALID_CREDENTIALS
                - TOKEN_EXPIRED
                - TOKEN_INVALID
                - USER_NOT_FOUND
                - USER_INACTIVE
                - EMAIL_NOT_VERIFIED
                - RATE_LIMIT_EXCEEDED
                - INSUFFICIENT_PERMISSIONS
                - VALIDATION_ERROR
                - INTERNAL_ERROR
                - SERVICE_UNAVAILABLE
                - AGENT_UNAUTHORIZED
            message:
              type: string
              description: Human-readable error message
            details:
              type: object
              description: Additional error details
            trace_id:
              type: string
              description: Request trace identifier for debugging
        meta:
          $ref: '#/components/schemas/ResponseMeta'

    # Success Response Models
    APIResponse:
      type: object
      required:
        - success
        - data
      properties:
        success:
          type: boolean
          enum: [true]
        data:
          type: object
          description: Response data
        error:
          type: object
          nullable: true
        meta:
          $ref: '#/components/schemas/ResponseMeta'
    ResponseMeta:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
        version:
          type: string
          description: API version
        request_id:
          type: string
          description: Unique request identifier
        rate_limit:
          $ref: '#/components/schemas/RateLimitInfo'
        performance:
          $ref: '#/components/schemas/PerformanceInfo'
    RateLimitInfo:
      type: object
      properties:
        limit:
          type: integer
          description: Rate limit threshold
        remaining:
          type: integer
          description: Remaining requests in window
        reset:
          type: string
          format: date-time
          description: When the rate limit resets
    PerformanceInfo:
      type: object
      properties:
        response_time_ms:
          type: number
          description: Response time in milliseconds
        cache_hit:
          type: boolean
          description: Whether response was served from cache

    # Pagination Models
...
  examples:
    LoginSuccess:
      summary: Successful login response
      value:
        success: true
        data:
          access_token: "eyJhbGciOiJSUzI1NiIsInR5cCI6IkpXVCJ9..."
          refresh_token: "rt_1234567890abcdef"
          token_type: "Bearer"
          expires_in: 3600
          user:
            id: "123e4567-e89b-12d3-a456-426614174000"
            email: "user@example.com"
            name: "John Doe"
            roles: ["user"]
            is_active: true
            email_verified: true
          session_id: "sess_1234567890"
        meta:
          timestamp: "2025-08-31T10:00:00Z"
          version: "2.0.0"
          request_id: "req_1234567890"
    ValidationError:
      summary: Validation error response
      value:
        success: false
        data: null
        error:
          code: "VALIDATION_ERROR"
          message: "Request validation failed"
          details:
            email: "Invalid email format"
            password: "Password too weak"
          trace_id: "trace_1234567890"
        meta:
          timestamp: "2025-08-31T10:00:00Z"
          version: "2.0.0"
          request_id: "req_1234567890"

# API Endpoints
paths:
...
  /auth/login:
    post:
      tags: [Authentication]
      summary: User login
      description: |
        Authenticate user credentials and return JWT tokens.
        
        **Performance**: Optimized for <50ms response time with multi-layered caching.
        **Security**: Rate limited to 5 attempts per minute per IP.
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              email: "user@example.com"
              password: "SecureP@ssw0rd"
              remember_me: false
      responses:
        200:
          description: Login successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
              examples:
                success:
                  $ref: '#/components/examples/LoginSuccess'
          headers:
            X-RateLimit-Limit:
              schema:
                type: integer
              description: Rate limit threshold
            X-RateLimit-Remaining:
              schema:
                type: integer
              description: Remaining requests
        400:
          description: Invalid request data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
              examples:
                validation_error:
                  $ref: '#/components/examples/ValidationError'
        401:
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
        429:
          description: Rate limit exceeded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
...
  /auth/logout:
    post:
      tags: [Authentication]
      summary: User logout
      description: |
        Invalidate current session and tokens.
        
        **Security**: Clears all authentication traces and adds tokens to blacklist.
      operationId: logout
      security:
        - BearerAuth: []
      responses:
        200:
          description: Logout successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIResponse'
        401:
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/APIError'
...