# scwt-test-repo/backend/auth_endpoint.py
"""
Auth endpoint stub for SCWT testing
TODO: Implement secure authentication with JWT tokens
"""
from fastapi import APIRouter, HTTPException
from fastapi.security import HTTPBearer

router = APIRouter()
security = HTTPBearer()

@router.post("/auth/login")
async def login(credentials: dict):
    # TODO: Implement authentication logic
    raise HTTPException(status_code=501, detail="Not implemented")

@router.post("/auth/logout") 
async def logout():
    # TODO: Implement logout logic
    raise HTTPException(status_code=501, detail="Not implemented")